Improve codes

+++ src/main/java/com/itersdesktop/javatechs/concurrent/ParallelForkJoinExample.java
@@ -9,7 +9,7 @@ public class ParallelForkJoinExample {
     public static void main(String[] args) {
         List<Integer> numbers = Arrays.asList(0, 1, 2, 3, 4, 5, 6, 7, 8, 9);
         int availProcessors = Runtime.getRuntime().availableProcessors();
-        int maxNbThread = availProcessors > 4 ? (int) availProcessors/2 : 4;
+        int maxNbThread = availProcessors > 4 ? availProcessors /2 : 4;
         ForkJoinPool customForkJoinPool = new ForkJoinPool(maxNbThread);
 
         try {
